[0002-11-consumer] refactor: (producer) reorganize packages and refactor code
