+++ producer/src/main/java/com/ppm/delivery/order/producer/exception/MessageErrorConstants.java
@@ -11,8 +11,7 @@ public class MessageErrorConstants {
     public static final String ERROR_PROFILE_REQUIRED_HEADER = "Profile is required in the request header";
     public static final String ERROR_PROFILE_IS_INVALID = "Profile is Invalid";
     public static final String ERROR_CORRELATION_ID_REQUIRED_HEADER = "CorrelationId is required in the request header";
-    // TODO atg - Review consumer - Renomear para Timestamp
-    public static final String ERROR_TIME_STAMP_REQUIRED_HEADER = "TimeStamp is required in the request header";
+    public static final String ERROR_TIMESTAMP_REQUIRED_HEADER = "Timestamp is required in the request header";
     public static final String ERROR_COUNTRY_AND_ACTION_MUST_NOT_BE_NULL_OR_EMPTY = "Country and action must not be null or empty.";
     public static final String ERROR_UNSUPPORTED_COUNTRY_ACTION_COMBINATION = "Unsupported country-action combination";
 }
