+++ producer/src/main/java/com/ppm/delivery/order/producer/message/constants/MessageHeaderConstants.java
@@ -1,4 +1,4 @@
-package com.ppm.delivery.order.producer.message.config;
+package com.ppm.delivery.order.producer.message.constants;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
